refactor: move 3D Skin layers mixins outside of epicfight.mixins.json
