+++ src/main/java/yesman/epicfight/compat/skinlayer3d/mixin/MixinCustomModelPart.java
@@ -1,4 +1,4 @@
-package yesman.epicfight.mixin.skinlayers;
+package yesman.epicfight.compat.skinlayer3d.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
