+++ src/main/resources/epicfight-compat.skinlayers3d.mixins.json
@@ -1,0 +1,13 @@
+{
+  "required": false,
+  "minVersion": "0.8",
+  "compatibilityLevel": "JAVA_21",
+  "package": "yesman.epicfight.compat.skinlayer3d.mixin",
+  "client": [
+    "MixinCustomModelPart",
+    "MixinCustomizableCubeWrapper$SkinLayer3DMixinCustomModelCube"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
