+++ src/main/resources/epicfight.mixins.json
@@ -24,8 +24,6 @@
 		"client.MixinParticleEngine",
 		"client.MixinThrownTridentRenderer",
 		"geckolib.MixinGeoArmorRenderer",
-		"skinlayers.MixinCustomModelPart",
-		"skinlayers.MixinCustomizableCubeWrapper$SkinLayer3DMixinCustomModelCube",
 		"teamlapen.MixinVampirePlayerHeadLayer",
 		"teamlapen.MixinHumanWerewolfLayer"
 	],
