+++ src/main/templates/META-INF/neoforge.mods.toml
@@ -19,6 +19,9 @@ config="${mod_id}.mixins.json"
 [[mixins]]
 config="${mod_id}-platform.neoforge.mixins.json"
 
+[[mixins]]
+config="${mod_id}-compat.skinlayers3d.mixins.json"
+
 [[accessTransformers]]
 file="META-INF/accesstransformer.cfg"
